requirements.txt
ors-sdk>=0.1.0
numpy>=1.24.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
openai>=1.30.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
matplotlib>=3.7.0
pandas>=2.0.0
httpx>=0.24.0
tenacity>=8.2.0
structlog>=23.0.0
sse-starlette>=1.6.0
aiohttp>=3.8.0
typing-extensions>=4.7.0
